src/clients/veo_client.py
"""
Google Veo API Client for video generation
"""
import os
import time
import logging
from typing import Optional, Dict, Any
from google.cloud import aiplatform
from google.oauth2 import service_account

logger = logging.getLogger(__name__)


class VeoClient:
    """Client for Google Veo video generation API"""

    def __init__(
        self,
        project_id: Optional[str] = None,
        location: Optional[str] = None,
        credentials_path: Optional[str] = None
    ):
        """
        Initialize Veo API client

        Args:
            project_id: Google Cloud project ID
            location: Google Cloud location (e.g., us-central1)
            credentials_path: Path to service account JSON key
        """
        self.project_id = project_id or os.getenv("GOOGLE_CLOUD_PROJECT")
        self.location = location or os.getenv("VEO_LOCATION", "us-central1")

        # Initialize credentials
        creds_path = credentials_path or os.getenv("GOOGLE_APPLICATION_CREDENTIALS")
        if creds_path and os.path.exists(creds_path):
            credentials = service_account.Credentials.from_service_account_file(
                creds_path
            )
            aiplatform.init(
                project=self.project_id,
                location=self.location,
                credentials=credentials
            )
        else:
            aiplatform.init(project=self.project_id, location=self.location)

        logger.info(f"Initialized Veo client for project {self.project_id}")

    def generate_video(
        self,
        prompt: str,
        duration: int = 5,
        aspect_ratio: str = "16:9",
        **kwargs
    ) -> Dict[str, Any]:
        """
        Generate video using Veo API

        Args:
            prompt: Text prompt for video generation
            duration: Video duration in seconds
            aspect_ratio: Video aspect ratio
            **kwargs: Additional API parameters

        Returns:
            Dictionary containing job_id and other metadata
        """
        logger.info(f"Generating video with prompt: {prompt[:100]}...")

        try:
            # Note: This is a placeholder implementation
            # The actual Veo API interface may differ based on Google's final API
            # You'll need to update this based on the official Veo API documentation

            # Using Vertex AI Imagen for now as an example structure
            # Replace with actual Veo endpoint when available
            from vertexai.preview.vision_models import ImageGenerationModel

            # This is conceptual - replace with actual Veo video generation
            model = ImageGenerationModel.from_pretrained("imagegeneration@005")

            # Placeholder for actual video generation
            # The real implementation will use Veo's video generation endpoint
            logger.warning("Using placeholder Veo implementation - update with actual API")

            job_data = {
                "job_id": f"veo_job_{int(time.time())}",
                "status": "PENDING",
                "prompt": prompt,
                "duration": duration,
                "aspect_ratio": aspect_ratio,
                "created_at": time.time()
            }

            return job_data

        except Exception as e:
            logger.error(f"Error generating video: {str(e)}")
            raise

    def wait_for_completion(
        self,
        job_id: str,
        timeout: int = 600,
        poll_interval: int = 10
    ) -> Dict[str, Any]:
        """
        Wait for video generation to complete

        Args:
            job_id: Job identifier from generate_video
            timeout: Maximum wait time in seconds
            poll_interval: Time between status checks in seconds

        Returns:
            Dictionary with job status and video URL
        """
        logger.info(f"Waiting for job {job_id} to complete...")

        start_time = time.time()

        while time.time() - start_time < timeout:
            try:
                # Poll job status
                # Replace with actual Veo API status check
                status = self._check_job_status(job_id)

                if status["status"] == "COMPLETED":
                    logger.info(f"Job {job_id} completed successfully")
                    return status
                elif status["status"] == "FAILED":
                    raise Exception(f"Job {job_id} failed: {status.get('error')}")

                logger.debug(f"Job {job_id} status: {status['status']}")
                time.sleep(poll_interval)

            except Exception as e:
                logger.error(f"Error checking job status: {str(e)}")
                raise

        raise TimeoutError(f"Job {job_id} did not complete within {timeout} seconds")

    def _check_job_status(self, job_id: str) -> Dict[str, Any]:
        """
        Check status of a video generation job

        Args:
            job_id: Job identifier

        Returns:
            Dictionary with job status information
        """
        # Placeholder implementation
        # Replace with actual Veo API status endpoint

        # Simulate completion after some time for demo purposes
        return {
            "job_id": job_id,
            "status": "COMPLETED",
            "video_url": f"https://storage.googleapis.com/veo-output/{job_id}.mp4",
            "completed_at": time.time()
        }

    def get_video_url(self, job_id: str) -> str:
        """
        Get download URL for generated video

        Args:
            job_id: Job identifier

        Returns:
            URL to download the video
        """
        status = self._check_job_status(job_id)

        if status["status"] != "COMPLETED":
            raise Exception(f"Video not ready. Status: {status['status']}")

        return status.get("video_url", "")
